separate s201224434 into AutoIncrementerTest, ConfigNewTest, UtilitesTest
